New Resource Library helping to manage textures, shaders, fonts, etc.
- Added a Libraries and Library classes to help manage textures,
shaders, meshes, fonts, etc.
- Added a method to both bind a texture and associate it to a render
target and to an uniform in the current shader.
- Corrected a bug in PlaneMesh (orientation of textures).

+++ src/main/scala/org/sofa/opengl/backend/SurfaceJogl.scala
@@ -146,9 +146,9 @@ class MotionEventAWT(source:AWTMouseEvent, pressed:Boolean, released:Boolean) ex
     def sourceEvent:AnyRef = source
 }
 
-// ========================================================================
-// == NEWT ================================================================
-// ========================================================================
+// =======================================================================================================================================
+// == NEWT ===============================================================================================================================
+// =======================================================================================================================================
 
 class SurfaceNewt(
     val renderer:SurfaceRenderer,
@@ -267,7 +267,7 @@ class KeyEventJogl(val source:JoglKeyEvent) extends KeyEvent {
 }
 
 class ScrollEventJogl(source:JoglMouseEvent) extends ScrollEvent {
-    def amount:Int = source.getWheelRotation
+    def amount:Int = source.getWheelRotation.toInt
 }
 
 class MotionEventJogl(source:JoglMouseEvent, pressed:Boolean, released:Boolean) extends MotionEvent {
